backend/requirements.txt
django==3.0.8
paho-mqtt==1.6.1
asgiref==3.5.0
djangorestframework==3.13.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
dj-database-url==0.5.0
markdown
django-filter


